backend/requirements.txt
# Core
Flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0

# Auth
PyJWT==2.8.0
bcrypt==4.1.2

# Database
mysql-connector-python==8.3.0

# ML
tensorflow==2.15.0
opencv-python-headless==4.9.0.80
numpy==1.26.4
Pillow==10.2.0
scikit-learn==1.4.0

# HTTP
requests==2.31.0

# Payment
razorpay==2.0.1

# Utilities
Werkzeug==3.0.1